Debug
debug: TTC字体内嵌应该只读取其中的目标TTF
debug: 不子集化的内嵌字体文件名不应该加_subset
improve: 提升字幕Fonts行读取性能
change: UI按钮宽度小改
trivial: DllExport.cpp不再抛出异常
trivial: version文件改名

+++ fontmatch_dll/fontmatch/FontMatch3.cpp
@@ -55,7 +55,7 @@ const wstring FontMatch3::GetMatchingFont(unordered_map<FONT_PROPERTY, wstring>&
     // 进行严格筛选 ------
     ComPtr<IDWriteFontSet> fontSet; // 筛选结果集合
     if (fontProps.empty()
-        || FAILED(this->systemFontSet->GetMatchingFonts(fontProps.data(), fontProps.size(), &fontSet))
+        || FAILED(this->systemFontSet->GetMatchingFonts(fontProps.data(), (UINT32)fontProps.size(), &fontSet))
         || fontSet->GetFontCount() == 0)    // 这个重载版本的GetMatchingFonts只能按照各种名字筛选，不能筛选weight和style
         return wstring();
